Add flexible NEM optimization period + sub-hourly resolution
The NEM single-day reference path can now solve any calendar month or
custom date range at 1h/30min/5min resolution, not just a fixed March
month at hourly. Adds nem_data.period_ts() (general period+resolution
window over the native 5-min SCADA/price cache) alongside the existing
reference_month_ts(), which is left untouched.

Fixes two latent resolution-hardcoding bugs this surfaced:
- extract_results() summed MW dispatch directly into "MWh"/revenue
  figures, only correct at exactly 1h/row -- would have overstated
  every volume/revenue number 2x/12x at 30min/5min resolution.
- compute_counterfactuals() hardcoded dt=1.0 hours/row.

Both now take an explicit resolution_h/dt, default 1.0 (unchanged
behavior for every existing hourly-only caller). ui/tabs/optimization.py
gets a period (month vs custom range) + resolution picker for NEM
scenarios, with a snapshot-count warning about solve time/memory on
constrained hosts.

+++ ppa/financials.py
@@ -46,7 +46,7 @@ def run_financial_analysis(
     scenario: Scenario,
     summary: SummaryVolumes,
     revenue: RevenueBreakdown,
-    n_period_hours: int,
+    n_period_hours: float,
 ) -> FinancialResult:
     s = scenario
     scale = HOURS_PER_YEAR / n_period_hours
